# Hustle-Safe/Hustle-Safe/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleSafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all workers
      parameters:
        - name: zone_id
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerListResponse"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a new worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerRequest"
      responses:
        "201":
          description: Worker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Worker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Worker not found
  /workers/phone/{phone}:
    get:
      operationId: getWorkerByPhone
      tags: [workers]
      summary: Get a worker by phone number
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Worker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Worker not found
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List policies
      parameters:
        - name: worker_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a new policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a policy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List all claims
      parameters:
        - name: worker_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: zone_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimListResponse"
  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get a claim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Claim details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
    patch:
      operationId: updateClaim
      tags: [claims]
      summary: Update claim status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClaimRequest"
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /disruptions:
    get:
      operationId: listDisruptions
      tags: [disruptions]
      summary: List disruption events
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: zone_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of disruption events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisruptionListResponse"
  /simulator/trigger:
    post:
      operationId: triggerDisruption
      tags: [simulator]
      summary: Trigger a disruption event in a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerDisruptionRequest"
      responses:
        "200":
          description: Disruption triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerDisruptionResponse"
  /simulator/resolve/{zoneId}:
    post:
      operationId: resolveDisruption
      tags: [simulator]
      summary: Resolve/end a disruption in a zone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disruption resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveDisruptionResponse"
...
components:
  schemas:
...
    Worker:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        platform:
          type: string
          enum: [zomato, swiggy, both]
        zone_id:
          type: string
        upi_id:
          type: string
          nullable: true
        platform_rating:
          type: number
        policy_tier:
          type: string
          enum: [basic, standard, pro]
        is_active:
          type: boolean
        fraud_score:
          type: number
        account_age_days:
          type: integer
        created_at:
          type: string
      required:
        - id
        - name
        - phone
        - platform
        - zone_id
        - platform_rating
        - policy_tier
        - is_active
        - fraud_score
        - account_age_days
        - created_at
    WorkerListResponse:
      type: object
      properties:
        workers:
          type: array
          items:
            $ref: "#/components/schemas/Worker"
        total:
          type: integer
      required:
        - workers
        - total
    CreateWorkerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        platform:
          type: string
          enum: [zomato, swiggy, both]
        zone_id:
          type: string
        upi_id:
          type: string
        policy_tier:
          type: string
          enum: [basic, standard, pro]
      required:
        - name
        - phone
        - platform
        - zone_id
        - policy_tier
    Zone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        gds_score:
          type: integer
        status:
          type: string
          enum: [normal, elevated, high, shutdown]
        active_workers:
          type: integer
        rainfall_mm:
          type: number
        traffic_score:
          type: number
        aqi:
          type: integer
        demand_drop_pct:
          type: integer
        govt_alert:
          type: boolean
        last_updated:
          type: string
      required:
        - id
        - name
        - city
        - gds_score
        - status
        - active_workers
        - rainfall_mm
        - traffic_score
        - aqi
        - demand_drop_pct
        - govt_alert
        - last_updated
...
    Policy:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        tier:
          type: string
          enum: [basic, standard, pro]
        weekly_premium:
          type: number
        coverage_cap:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
        activated_at:
          type: string
        expires_at:
          type: string
          nullable: true
        zone_id:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - worker_id
        - tier
        - weekly_premium
        - coverage_cap
        - status
        - activated_at
        - created_at
    PolicyListResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
        total:
          type: integer
      required:
        - policies
        - total
    CreatePolicyRequest:
      type: object
      properties:
        worker_id:
          type: string
        tier:
          type: string
          enum: [basic, standard, pro]
        zone_id:
          type: string
      required:
        - worker_id
        - tier
    Claim:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        policy_id:
          type: string
          nullable: true
        zone_id:
          type: string
          nullable: true
        disruption_type:
          type: string
          nullable: true
        disruption_start:
          type: string
          nullable: true
        disruption_end:
          type: string
          nullable: true
        hours_affected:
          type: number
          nullable: true
        hourly_rate:
          type: number
        payout_amount:
          type: number
          nullable: true
        fraud_score:
          type: number
        status:
          type: string
          enum: [pending, auto_approved, soft_hold, insurer_review, auto_rejected, paid]
        fraud_signals:
          type: object
          additionalProperties: true
        resolution_notes:
          type: string
          nullable: true
        created_at:
          type: string
        paid_at:
          type: string
          nullable: true
        worker_name:
          type: string
          nullable: true
        zone_name:
          type: string
          nullable: true
      required:
        - id
        - worker_id
        - hourly_rate
        - fraud_score
        - status
        - fraud_signals
        - created_at
    ClaimListResponse:
      type: object
      properties:
        claims:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        total:
          type: integer
      required:
        - claims
        - total
    UpdateClaimRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, auto_approved, soft_hold, insurer_review, auto_rejected, paid]
        resolution_notes:
          type: string
    DisruptionEvent:
      type: object
      properties:
        id:
          type: string
        zone_id:
          type: string
          nullable: true
        event_type:
          type: string
          enum: [heavy_rain, flood, curfew, platform_outage, strike, aqi_hazard]
        gds_target:
          type: integer
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        triggered_by:
          type: string
        is_active:
          type: boolean
        started_at:
          type: string
          nullable: true
        ended_at:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - event_type
        - triggered_by
        - is_active
        - created_at
    DisruptionListResponse:
      type: object
      properties:
        disruptions:
          type: array
          items:
            $ref: "#/components/schemas/DisruptionEvent"
      required:
        - disruptions
    TriggerDisruptionRequest:
      type: object
      properties:
        zone_id:
          type: string
        event_type:
          type: string
          enum: [heavy_rain, flood, curfew, platform_outage, strike, aqi_hazard]
        gds_target:
          type: integer
        duration_minutes:
          type: integer
      required:
        - zone_id
        - event_type
        - gds_target
    TriggerDisruptionResponse:
      type: object
      properties:
        success:
          type: boolean
        disruption:
          $ref: "#/components/schemas/DisruptionEvent"
        zone:
          $ref: "#/components/schemas/Zone"
        claims_created:
          type: integer
        message:
          type: string
      required:
        - success
        - disruption
        - zone
        - claims_created
        - message
    ResolveDisruptionResponse:
      type: object
      properties:
        success:
          type: boolean
        zone:
          $ref: "#/components/schemas/Zone"
        message:
          type: string
      required:
        - success
        - zone
        - message
...