# Hustle-Safe/Hustle-Safe/lib/api-zod/src/generated/types/claim.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HustleSafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClaimFraudSignals } from "./claimFraudSignals";
import type { ClaimStatus } from "./claimStatus";

export interface Claim {
  id: string;
  worker_id: string;
  policy_id?: string | null;
  zone_id?: string | null;
  disruption_type?: string | null;
  disruption_start?: string | null;
  disruption_end?: string | null;
  hours_affected?: number | null;
  hourly_rate: number;
  payout_amount?: number | null;
  fraud_score: number;
  status: ClaimStatus;
  fraud_signals: ClaimFraudSignals;
  resolution_notes?: string | null;
  created_at: string;
  paid_at?: string | null;
  worker_name?: string | null;
  zone_name?: string | null;
}