# Hustle-Safe/Hustle-Safe/lib/api-zod/src/generated/types/policyListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HustleSafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Policy } from "./policy";

export interface PolicyListResponse {
  policies: Policy[];
  total: number;
}