# Hustle-Safe/Hustle-Safe/lib/api-zod/src/generated/types/triggerDisruptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HustleSafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TriggerDisruptionRequestEventType } from "./triggerDisruptionRequestEventType";

export interface TriggerDisruptionRequest {
  zone_id: string;
  event_type: TriggerDisruptionRequestEventType;
  gds_target: number;
  duration_minutes?: number;
}