# Hustle-Safe/Hustle-Safe/lib/api-zod/src/generated/types/worker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HustleSafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkerPlatform } from "./workerPlatform";
import type { WorkerPolicyTier } from "./workerPolicyTier";

export interface Worker {
  id: string;
  name: string;
  phone: string;
  email?: string | null;
  platform: WorkerPlatform;
  zone_id: string;
  upi_id?: string | null;
  platform_rating: number;
  policy_tier: WorkerPolicyTier;
  is_active: boolean;
  fraud_score: number;
  account_age_days: number;
  created_at: string;
}